Break StateManager to it's own file
