+++ src/drivers/serac.cpp
@@ -29,6 +29,7 @@
 #include "serac/numerics/mesh_utils.hpp"
 #include "serac/physics/thermal_solid.hpp"
 #include "serac/physics/utilities/equation_solver.hpp"
+#include "serac/physics/utilities/state_manager.hpp"
 #include "serac/serac_config.hpp"
 
 namespace serac {
