+++ src/serac/physics/utilities/CMakeLists.txt
@@ -13,6 +13,7 @@ set(physics_utilities_headers
     finite_element_state.hpp
     physics_utils.hpp
     solver_config.hpp
+    state_manager.hpp
     )
 
 set(physics_utilities_sources
@@ -21,6 +22,7 @@ set(physics_utilities_sources
     equation_solver.cpp
     finite_element_state.cpp
     physics_utils.cpp
+    state_manager.cpp
     )
 
 set(physics_utilities_depends serac_infrastructure)
