+++ tests/serac_solid_reuse_mesh.cpp
@@ -14,6 +14,7 @@
 #include "serac/coefficients/coefficient_extensions.hpp"
 #include "serac/infrastructure/input.hpp"
 #include "serac/numerics/mesh_utils.hpp"
+#include "serac/physics/utilities/state_manager.hpp"
 #include "serac/serac_config.hpp"
 #include "test_utilities.hpp"
