+++ tests/serac_thermal_solver.cpp
@@ -14,6 +14,7 @@
 #include "mfem.hpp"
 
 #include "serac/numerics/mesh_utils.hpp"
+#include "serac/physics/utilities/state_manager.hpp"
 #include "serac/serac_config.hpp"
 #include "test_utilities.hpp"
