+++ tests/serac_thermal_structural_solver.cpp
@@ -13,6 +13,7 @@
 
 #include "serac/coefficients/coefficient_extensions.hpp"
 #include "serac/numerics/mesh_utils.hpp"
+#include "serac/physics/utilities/state_manager.hpp"
 #include "serac/serac_config.hpp"
 
 namespace serac {
