+++ tests/test_utilities.cpp
@@ -12,6 +12,7 @@
 #include "serac/numerics/mesh_utils.hpp"
 #include "serac/physics/solid.hpp"
 #include "serac/physics/thermal_conduction.hpp"
+#include "serac/physics/utilities/state_manager.hpp"
 
 namespace serac {
